che-test/scripts/autotests/requirements.txt
asn1crypto
bcrypt
certifi
cffi
chardet
cryptography
dnspython
dredd-hooks
idna
jsonpickle
jsonschema
mysql-connector-python
paramiko
pyasn1
pycparser
PyNaCl
python-dateutil
pytest
PyYAML
requests
selenium
six
temp-mail
